+++ packages/terra-clinical-item-display/CHANGELOG.md
@@ -7,7 +7,11 @@ Unreleased
 4.0.0 - (June 23, 2020)
 ----------
 ### Breaking
+* Removed `ItemDisplay.Comment`.
 * Changed the text style to primary by default and updated Jest snapshots.
+
+### Changed
+* Resolved dependency cycle.
 
 3.18.0 - (June 9, 2020)
 ----------